+++ vtm-gdx/src/org/oscim/gdx/InputHandler.java
@@ -1,5 +1,7 @@
 /*
+ * Copyright 2013 Hannes Janetzek
  * Copyright 2016-2017 devemux86
+ * Copyright 2017 Longri
  *
  * This file is part of the OpenScienceMap project (http://www.opensciencemap.org).
  *
@@ -22,11 +24,16 @@ import com.badlogic.gdx.Input.Buttons;
 import com.badlogic.gdx.InputProcessor;
 
 import org.oscim.layers.GenericLayer;
+import org.oscim.layers.GroupLayer;
+import org.oscim.layers.Layer;
 import org.oscim.layers.TileGridLayer;
+import org.oscim.layers.tile.buildings.BuildingLayer;
 import org.oscim.map.Map;
 import org.oscim.map.ViewController;
 import org.oscim.theme.VtmThemes;
 import org.oscim.utils.Easing;
+
+import java.util.List;
 
 public class InputHandler implements InputProcessor {
 
@@ -145,6 +152,11 @@ public class InputHandler implements InputProcessor {
                 }
                 mMap.render();
                 break;
+
+            case Input.Keys.B:
+                toggleBuildingLayer(mMap.layers());
+                mMap.render();
+                break;
         }
         return false;
     }
@@ -241,5 +253,19 @@ public class InputHandler implements InputProcessor {
         mMap.animator().animateZoom(250, amount > 0 ? 0.75f : 1.333f, fx, fy, Easing.Type.LINEAR);
         mMap.updateMap(false);
         return true;
+    }
+
+    private boolean toggleBuildingLayer(List<Layer> layers) {
+        for (Layer layer : layers) {
+            if (layer instanceof BuildingLayer) {
+                layer.setEnabled(!layer.isEnabled());
+                return true;
+            } else if (layer instanceof GroupLayer) {
+                if (toggleBuildingLayer(((GroupLayer) layer).layers)) {
+                    return true;
+                }
+            }
+        }
+        return false;
     }
 }