+++ vtm/src/org/oscim/layers/tile/vector/labeling/LabelLayer.java
@@ -1,5 +1,8 @@
 /*
  * Copyright 2012, 2013 Hannes Janetzek
+ * Copyright 2017 Wolfgang Schramm
+ * Copyright 2017 devemux86
+ * Copyright 2017 Andrey Novikov
  *
  * This file is part of the OpenScienceMap project (http://www.opensciencemap.org).
  *
@@ -39,9 +42,13 @@ public class LabelLayer extends Layer implements Map.UpdateListener, TileManager
     private final Worker mWorker;
 
     public LabelLayer(Map map, VectorTileLayer l) {
+        this(map, l, new LabelTileLoaderHook());
+    }
+
+    public LabelLayer(Map map, VectorTileLayer l, VectorTileLayer.TileLoaderThemeHook h) {
         super(map);
         l.getManager().events.bind(this);
-        l.addHook(new LabelTileLoaderHook());
+        l.addHook(h);
 
         mLabelPlacer = new LabelPlacement(map, l.tileRenderer());
         mWorker = new Worker(map);
@@ -84,6 +91,9 @@ public class LabelLayer extends Layer implements Map.UpdateListener, TileManager
     }
 
     public void update() {
+        if (!isEnabled())
+            return;
+
         mWorker.submit(MAX_RELABEL_DELAY);
     }
 
@@ -98,6 +108,9 @@ public class LabelLayer extends Layer implements Map.UpdateListener, TileManager
 
         if (event == Map.CLEAR_EVENT)
             mWorker.cancel(true);
+
+        if (!isEnabled())
+            return;
 
         if (event == Map.POSITION_EVENT)
             mWorker.submit(MAX_RELABEL_DELAY);
@@ -123,7 +136,7 @@ public class LabelLayer extends Layer implements Map.UpdateListener, TileManager
     @Override
     public void onTileManagerEvent(Event e, MapTile tile) {
         if (e == TileManager.TILE_LOADED) {
-            if (tile.isVisible)
+            if (tile.isVisible && isEnabled())
                 mWorker.submit(MAX_RELABEL_DELAY / 4);
             //log.debug("tile loaded: {}", tile);
         } else if (e == TileManager.TILE_REMOVED) {