+++ vtm/src/org/oscim/renderer/LocationRenderer.java
@@ -18,7 +18,9 @@
  */
 package org.oscim.renderer;
 
+import org.oscim.backend.CanvasAdapter;
 import org.oscim.backend.GL;
+import org.oscim.backend.canvas.Color;
 import org.oscim.core.Box;
 import org.oscim.core.Point;
 import org.oscim.core.Tile;
@@ -34,11 +36,13 @@ public class LocationRenderer extends LayerRenderer {
     private static final long ANIM_RATE = 50;
     private static final long INTERVAL = 2000;
 
-    private static final float CIRCLE_SIZE = 60;
+    private static final float CIRCLE_SIZE = 30;
+    private static final int COLOR = 0xff3333cc;
     private static final int SHOW_ACCURACY_ZOOM = 16;
 
     private final Map mMap;
     private final Layer mLayer;
+    private final float mScale;
 
     private String mShaderFile;
     private int mShaderProgram;
@@ -47,6 +51,7 @@ public class LocationRenderer extends LayerRenderer {
     private int hScale;
     private int hPhase;
     private int hDirection;
+    private int uColor;
     private int uMode;
 
     private final Point mIndicatorPosition = new Point();
@@ -62,17 +67,37 @@ public class LocationRenderer extends LayerRenderer {
     private long mAnimStart;
 
     private Callback mCallback;
+    private final float[] mColors = new float[4];
     private final Point mLocation = new Point(Double.NaN, Double.NaN);
     private double mRadius;
     private int mShowAccuracyZoom = SHOW_ACCURACY_ZOOM;
 
     public LocationRenderer(Map map, Layer layer) {
+        this(map, layer, CanvasAdapter.dpi / CanvasAdapter.DEFAULT_DPI);
+    }
+
+    public LocationRenderer(Map map, Layer layer, float scale) {
         mMap = map;
         mLayer = layer;
+        mScale = scale;
+
+        float a = Color.aToFloat(COLOR);
+        mColors[0] = a * Color.rToFloat(COLOR);
+        mColors[1] = a * Color.gToFloat(COLOR);
+        mColors[2] = a * Color.bToFloat(COLOR);
+        mColors[3] = a;
     }
 
     public void setCallback(Callback callback) {
         mCallback = callback;
+    }
+
+    public void setColor(int color) {
+        float a = Color.aToFloat(color);
+        mColors[0] = a * Color.rToFloat(color);
+        mColors[1] = a * Color.gToFloat(color);
+        mColors[2] = a * Color.bToFloat(color);
+        mColors[3] = a;
     }
 
     public void setLocation(double x, double y, double radius) {
@@ -83,6 +108,7 @@ public class LocationRenderer extends LayerRenderer {
 
     public void setShader(String shaderFile) {
         mShaderFile = shaderFile;
+        mInitialized = false;
     }
 
     public void setShowAccuracyZoom(int showAccuracyZoom) {
@@ -107,7 +133,8 @@ public class LocationRenderer extends LayerRenderer {
 
                 long diff = System.currentTimeMillis() - lastRun;
                 mMap.postDelayed(this, Math.min(ANIM_RATE, diff));
-                mMap.render();
+                if (!mLocationIsVisible)
+                    mMap.render();
                 lastRun = System.currentTimeMillis();
             }
         };
@@ -193,7 +220,7 @@ public class LocationRenderer extends LayerRenderer {
         GLState.enableVertexArrays(hVertexPosition, -1);
         MapRenderer.bindQuadVertexVBO(hVertexPosition/*, true*/);
 
-        float radius = CIRCLE_SIZE;
+        float radius = CIRCLE_SIZE * mScale;
 
         animate(true);
         boolean viewShed = false;
@@ -202,7 +229,7 @@ public class LocationRenderer extends LayerRenderer {
         } else {
             if (v.pos.zoomLevel >= mShowAccuracyZoom)
                 radius = (float) (mRadius * v.pos.scale);
-            radius = Math.max(CIRCLE_SIZE, radius);
+            radius = Math.max(CIRCLE_SIZE * mScale, radius);
 
             viewShed = true;
             //animate(false);
@@ -234,11 +261,15 @@ public class LocationRenderer extends LayerRenderer {
                 gl.uniform2f(hDirection,
                         (float) Math.cos(Math.toRadians(rotation)),
                         (float) Math.sin(Math.toRadians(rotation)));
-                gl.uniform1i(uMode, 3); // With bearing
-            } else
-                gl.uniform1i(uMode, 2); // Without bearing
+                gl.uniform1i(uMode, 1); // With bearing
+            } else {
+                gl.uniform2f(hDirection, 0, 0);
+                gl.uniform1i(uMode, 0); // Without bearing
+            }
         } else
-            gl.uniform1i(uMode, 1); // Outside screen
+            gl.uniform1i(uMode, -1); // Outside screen
+
+        GLUtils.glUniform4fv(uColor, 1, mColors);
 
         gl.drawArrays(GL.TRIANGLE_STRIP, 0, 4);
     }
@@ -254,6 +285,7 @@ public class LocationRenderer extends LayerRenderer {
         hPhase = gl.getUniformLocation(program, "u_phase");
         hScale = gl.getUniformLocation(program, "u_scale");
         hDirection = gl.getUniformLocation(program, "u_dir");
+        uColor = gl.getUniformLocation(program, "u_color");
         uMode = gl.getUniformLocation(program, "u_mode");
 
         return true;